Moved settings loading/saving/updating to mainwindow and use events for it
