+++ events.py
@@ -5,3 +5,4 @@ UpdateGaugeEvent, EVT_UPDATE_GAUGE_EVENT = NewEvent()
 MessageEvent, EVT_MESSAGE_EVENT = NewEvent()
 AssignPartEvent, EVT_ASSIGN_PART_EVENT = NewEvent()
 PopulateFootprintListEvent, EVT_POPULATE_FOOTPRINT_LIST_EVENT = NewEvent()
+UpdateSetting, EVT_UPDATE_SETTING = NewEvent()
